core(fr): extract navigation errors from gather-runner
