+++ lighthouse-core/test/gather/gather-runner-test.js
@@ -14,7 +14,6 @@ const GatherRunner_ = require('../../gather/gather-runner.js');
 const assert = require('assert').strict;
 const Config = require('../../config/config.js');
 const unresolvedPerfLog = require('./../fixtures/unresolved-perflog.json');
-const NetworkRequest = require('../../lib/network-request.js');
 const LHError = require('../../lib/lh-error.js');
 const networkRecordsToDevtoolsLog = require('../network-records-to-devtools-log.js');
 const Driver = require('../../gather/driver.js');
@@ -31,10 +30,6 @@ const GatherRunner = {
   beginRecording: makeParamsOptional(GatherRunner_.beginRecording),
   collectArtifacts: makeParamsOptional(GatherRunner_.collectArtifacts),
   endRecording: makeParamsOptional(GatherRunner_.endRecording),
-  getInterstitialError: makeParamsOptional(GatherRunner_.getInterstitialError),
-  getNetworkError: makeParamsOptional(GatherRunner_.getNetworkError),
-  getNonHtmlError: makeParamsOptional(GatherRunner_.getNonHtmlError),
-  getPageLoadError: makeParamsOptional(GatherRunner_.getPageLoadError),
   getSlowHostCpuWarning: makeParamsOptional(GatherRunner_.getSlowHostCpuWarning),
   initializeBaseArtifacts: makeParamsOptional(GatherRunner_.initializeBaseArtifacts),
   loadPage: makeParamsOptional(GatherRunner_.loadPage),
@@ -58,12 +53,6 @@ function makeConfig(json) {
   }
   return /** @type {Config & {passes: Array<LH.Config.Pass>}} */ (config);
 }
-
-const LoadFailureMode = {
-  fatal: /** @type {'fatal'} */ ('fatal'),
-  ignore: /** @type {'ignore'} */ ('ignore'),
-  warn: /** @type {'warn'} */ ('warn'),
-};
 
 class TestGatherer extends Gatherer {
   constructor() {
@@ -413,7 +402,7 @@ describe('GatherRunner', function() {
   it('prepares target for navigation', async () => {
     const passConfig = {
       passName: 'default',
-      loadFailureMode: LoadFailureMode.ignore,
+      loadFailureMode: /** @type {'ignore'} */ ('ignore'),
       recordTrace: true,
       useThrottling: true,
       gatherers: [],
@@ -831,384 +820,6 @@ describe('GatherRunner', function() {
     // firstPass has a saved trace and devtoolsLog, secondPass has an error trace and log.
     expect(Object.keys(artifacts.traces)).toEqual(['firstPass', 'pageLoadError-secondPass']);
     expect(Object.keys(artifacts.devtoolsLogs)).toEqual(['firstPass', 'pageLoadError-secondPass']);
-  });
-
-  describe('#getNetworkError', () => {
-    /**
-     * @param {NetworkRequest=} mainRecord
-     */
-    function getAndExpectError(mainRecord) {
-      const error = GatherRunner.getNetworkError(mainRecord);
-      if (!error) throw new Error('expected a network error');
-      return error;
-    }
-
-    it('passes when the page is loaded', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      assert.ok(!GatherRunner.getNetworkError(mainRecord));
-    });
-
-    it('fails when page fails to load', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.failed = true;
-      mainRecord.localizedFailDescription = 'foobar';
-      const error = getAndExpectError(mainRecord);
-      assert.equal(error.message, 'FAILED_DOCUMENT_REQUEST');
-      assert.equal(error.code, 'FAILED_DOCUMENT_REQUEST');
-      expect(error.friendlyMessage)
-        .toBeDisplayString(/^Lighthouse was unable to reliably load.*foobar/);
-    });
-
-    it('fails when page times out', () => {
-      const error = getAndExpectError(undefined);
-      assert.equal(error.message, 'NO_DOCUMENT_REQUEST');
-      assert.equal(error.code, 'NO_DOCUMENT_REQUEST');
-      expect(error.friendlyMessage).toBeDisplayString(/^Lighthouse was unable to reliably load/);
-    });
-
-    it('fails when page returns with a 404', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.statusCode = 404;
-      const error = getAndExpectError(mainRecord);
-      assert.equal(error.message, 'ERRORED_DOCUMENT_REQUEST');
-      assert.equal(error.code, 'ERRORED_DOCUMENT_REQUEST');
-      expect(error.friendlyMessage)
-        .toBeDisplayString(/^Lighthouse was unable to reliably load.*404/);
-    });
-
-    it('fails when page returns with a 500', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.statusCode = 500;
-      const error = getAndExpectError(mainRecord);
-      assert.equal(error.message, 'ERRORED_DOCUMENT_REQUEST');
-      assert.equal(error.code, 'ERRORED_DOCUMENT_REQUEST');
-      expect(error.friendlyMessage)
-        .toBeDisplayString(/^Lighthouse was unable to reliably load.*500/);
-    });
-
-    it('fails when page domain doesn\'t resolve', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.failed = true;
-      mainRecord.localizedFailDescription = 'net::ERR_NAME_NOT_RESOLVED';
-      const error = getAndExpectError(mainRecord);
-      assert.equal(error.message, 'DNS_FAILURE');
-      assert.equal(error.code, 'DNS_FAILURE');
-      expect(error.friendlyMessage).toBeDisplayString(/^DNS servers could not resolve/);
-    });
-  });
-
-  describe('#getInterstitialError', () => {
-    /**
-     * @param {NetworkRequest} mainRecord
-     * @param {NetworkRequest[]} networkRecords
-     */
-    function getAndExpectError(mainRecord, networkRecords) {
-      const error = GatherRunner.getInterstitialError(mainRecord, networkRecords);
-      if (!error) throw new Error('expected an interstitial error');
-      return error;
-    }
-
-    it('passes when the page was not requested', () => {
-      expect(GatherRunner.getInterstitialError(undefined, [])).toBeUndefined();
-    });
-
-    it('passes when the page is loaded', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      expect(GatherRunner.getInterstitialError(mainRecord, [mainRecord])).toBeUndefined();
-    });
-
-    it('passes when page fails to load normally', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.failed = true;
-      mainRecord.localizedFailDescription = 'foobar';
-      expect(GatherRunner.getInterstitialError(mainRecord, [mainRecord])).toBeUndefined();
-    });
-
-    it('passes when page gets a generic interstitial but somehow also loads everything', () => {
-      // This case, AFAIK, is impossible, but we'll err on the side of not tanking the run.
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      const interstitialRecord = new NetworkRequest();
-      interstitialRecord.url = 'data:text/html;base64,abcdef';
-      interstitialRecord.documentURL = 'chrome-error://chromewebdata/';
-      const records = [mainRecord, interstitialRecord];
-      expect(GatherRunner.getInterstitialError(mainRecord, records)).toBeUndefined();
-    });
-
-    it('fails when page gets a generic interstitial', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.failed = true;
-      mainRecord.localizedFailDescription = 'ERR_CONNECTION_RESET';
-      const interstitialRecord = new NetworkRequest();
-      interstitialRecord.url = 'data:text/html;base64,abcdef';
-      interstitialRecord.documentURL = 'chrome-error://chromewebdata/';
-      const records = [mainRecord, interstitialRecord];
-      const error = getAndExpectError(mainRecord, records);
-      expect(error.message).toEqual('CHROME_INTERSTITIAL_ERROR');
-      expect(error.code).toEqual('CHROME_INTERSTITIAL_ERROR');
-      expect(error.friendlyMessage).toBeDisplayString(/^Chrome prevented/);
-    });
-
-    it('fails when page gets a security interstitial', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.failed = true;
-      mainRecord.localizedFailDescription = 'net::ERR_CERT_COMMON_NAME_INVALID';
-      const interstitialRecord = new NetworkRequest();
-      interstitialRecord.url = 'data:text/html;base64,abcdef';
-      interstitialRecord.documentURL = 'chrome-error://chromewebdata/';
-      const records = [mainRecord, interstitialRecord];
-      const error = getAndExpectError(mainRecord, records);
-      expect(error.message).toEqual('INSECURE_DOCUMENT_REQUEST');
-      expect(error.code).toEqual('INSECURE_DOCUMENT_REQUEST');
-      expect(error.friendlyMessage).toBeDisplayString(/valid security certificate/);
-      expect(error.friendlyMessage).toBeDisplayString(/net::ERR_CERT_COMMON_NAME_INVALID/);
-    });
-
-    it('passes when page iframe gets a generic interstitial', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.failed = false;
-      const iframeRecord = new NetworkRequest();
-      iframeRecord.failed = true;
-      iframeRecord.url = 'https://the-ad.com';
-      iframeRecord.documentURL = 'https://the-ad.com';
-      const interstitialRecord = new NetworkRequest();
-      interstitialRecord.url = 'data:text/html;base64,abcdef';
-      interstitialRecord.documentURL = 'chrome-error://chromewebdata/';
-      const records = [mainRecord, iframeRecord, interstitialRecord];
-      const error = GatherRunner.getInterstitialError(mainRecord, records);
-      expect(error).toBeUndefined();
-    });
-  });
-
-  describe('#getNonHtmlError', () => {
-    /**
-     * @param {NetworkRequest} mainRecord
-     */
-    function getAndExpectError(mainRecord) {
-      const error = GatherRunner.getNonHtmlError(mainRecord);
-      if (!error) throw new Error('expected a non-HTML error');
-      return error;
-    }
-
-    it('passes when the page was not requested', () => {
-      expect(GatherRunner.getNonHtmlError(undefined)).toBeUndefined();
-    });
-
-    it('passes when the page is of MIME type text/html', () => {
-      const url = 'http://the-page.com';
-      const mainRecord = new NetworkRequest();
-      const mimeType = 'text/html';
-      mainRecord.url = url;
-      mainRecord.mimeType = mimeType;
-      expect(GatherRunner.getNonHtmlError(mainRecord)).toBeUndefined();
-    });
-
-    it('fails when the page is not of MIME type text/html', () => {
-      const url = 'http://the-page.com';
-      const mimeType = 'application/xml';
-      const mainRecord = new NetworkRequest();
-      mainRecord.url = url;
-      mainRecord.mimeType = mimeType;
-      const error = getAndExpectError(mainRecord);
-      expect(error.message).toEqual('NOT_HTML');
-      expect(error.code).toEqual('NOT_HTML');
-      expect(error.friendlyMessage).toBeDisplayString(/is not HTML \(served as/);
-    });
-  });
-
-
-  describe('#getPageLoadError', () => {
-    /**
-     * @param {RecursivePartial<LH.Gatherer.PassContext>} passContext
-     * @param {RecursivePartial<LH.Gatherer.LoadData>} loadData
-     * @param {LH.LighthouseError|undefined} navigationError
-     */
-    function getAndExpectError(passContext, loadData, navigationError) {
-      const error = GatherRunner.getPageLoadError(passContext, loadData, navigationError);
-      if (!error) throw new Error('expected a page load error');
-      return error;
-    }
-
-    /** @type {LH.LighthouseError} */
-    let navigationError;
-
-    beforeEach(() => {
-      navigationError = /** @type {LH.LighthouseError} */ (new Error('NAVIGATION_ERROR'));
-    });
-
-    it('passes when the page is loaded', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.fatal},
-      };
-      const mainRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-      mainRecord.url = passContext.url;
-      mainRecord.mimeType = 'text/html';
-      const error = GatherRunner.getPageLoadError(passContext, loadData, undefined);
-      expect(error).toBeUndefined();
-    });
-
-    it('passes when the page is loaded, ignoring any fragment', () => {
-      const passContext = {
-        url: 'http://example.com/#/page/list',
-        passConfig: {loadFailureMode: LoadFailureMode.fatal},
-      };
-      const mainRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-      mainRecord.url = 'http://example.com';
-      mainRecord.mimeType = 'text/html';
-      const error = GatherRunner.getPageLoadError(passContext, loadData, undefined);
-      expect(error).toBeUndefined();
-    });
-
-    it('passes when the page is expected to fail', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.ignore},
-      };
-      const mainRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-      mainRecord.url = passContext.url;
-      mainRecord.failed = true;
-
-      const error = GatherRunner.getPageLoadError(passContext, loadData, undefined);
-      expect(error).toBeUndefined();
-    });
-
-    it('passes when the page redirects to MIME type text/html', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.fatal},
-      };
-      const mainRecord = new NetworkRequest();
-      const finalRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-
-      mainRecord.url = passContext.url;
-      mainRecord.redirectDestination = finalRecord;
-      finalRecord.url = 'http://the-redirected-page.com';
-      finalRecord.mimeType = 'text/html';
-
-      const error = GatherRunner.getPageLoadError(passContext, loadData, undefined);
-      expect(error).toBeUndefined();
-    });
-
-    it('fails with interstitial error first', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.fatal},
-      };
-      const mainRecord = new NetworkRequest();
-      const interstitialRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord, interstitialRecord]};
-
-      mainRecord.url = passContext.url;
-      mainRecord.failed = true;
-      interstitialRecord.url = 'data:text/html;base64,abcdef';
-      interstitialRecord.documentURL = 'chrome-error://chromewebdata/';
-
-      const error = getAndExpectError(passContext, loadData, navigationError);
-      expect(error.message).toEqual('CHROME_INTERSTITIAL_ERROR');
-    });
-
-    it('fails with network error second', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.fatal},
-      };
-      const mainRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-
-      mainRecord.url = passContext.url;
-      mainRecord.failed = true;
-
-      const error = getAndExpectError(passContext, loadData, navigationError);
-      expect(error.message).toEqual('FAILED_DOCUMENT_REQUEST');
-    });
-
-    it('fails with non-HTML error third', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.fatal},
-      };
-      const mainRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-
-      mainRecord.url = passContext.url;
-      mainRecord.mimeType = 'application/xml';
-
-      const error = getAndExpectError(passContext, loadData, navigationError);
-      expect(error.message).toEqual('NOT_HTML');
-    });
-
-    it('fails with nav error last', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.fatal},
-      };
-      const mainRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-
-      mainRecord.url = passContext.url;
-      mainRecord.mimeType = 'text/html';
-
-      const error = getAndExpectError(passContext, loadData, navigationError);
-      expect(error.message).toEqual('NAVIGATION_ERROR');
-    });
-
-    it('fails when loadFailureMode is warn', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.warn},
-      };
-      const mainRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-
-      mainRecord.url = passContext.url;
-      mainRecord.mimeType = 'text/html';
-
-      const error = getAndExpectError(passContext, loadData, navigationError);
-      expect(error.message).toEqual('NAVIGATION_ERROR');
-    });
-
-    it('fails with non-HTML when redirect is not HTML', () => {
-      const passContext = {
-        url: 'http://the-page.com',
-        passConfig: {loadFailureMode: LoadFailureMode.fatal},
-      };
-      const mainRecord = new NetworkRequest();
-      const finalRecord = new NetworkRequest();
-      const loadData = {networkRecords: [mainRecord]};
-
-      mainRecord.url = passContext.url;
-      mainRecord.redirectDestination = finalRecord;
-      finalRecord.url = 'http://the-redirected-page.com';
-
-      const error = getAndExpectError(passContext, loadData, navigationError);
-      expect(error.message).toEqual('NOT_HTML');
-    });
   });
 
   describe('artifact collection', () => {
